Development into Master

+++ dataloader/clean_database.py
@@ -7,7 +7,7 @@ from arango import (
 from arango.database import StandardDatabase
 from tqdm import tqdm as tqdm
 
-from global_search_function import clean_analyzers
+from global_search import clean_analyzers
 
 from dataloader_constants import (
     COLLECTION_NAMES,
@@ -20,18 +20,13 @@ from dataloader_constants import (
     COLLECTION_FILES,
     COLLECTION_MENU_COLLECTIONS,
     COLLECTION_MENU_CATEGORIES,
-    COLLECTION_FILES_PARALLEL_COUNT,
-    EDGE_COLLECTION_NAMES,
-    COLLECTION_CATEGORIES_PARALLEL_COUNT,
-    EDGE_COLLECTION_COLLECTION_HAS_CATEGORIES,
-    GRAPH_COLLECTIONS_CATEGORIES,
     COLLECTION_LANGUAGES,
-    EDGE_COLLECTION_LANGUAGE_HAS_COLLECTIONS,
-    EDGE_COLLECTION_CATEGORY_HAS_FILES,
-
+    GLOBAL_STATS_CATEGORIES,
+    GLOBAL_STATS_FILES,
 )
 
-from dataloader_utils import get_database, get_system_database
+from utils import get_database, get_system_database
+
 
 def clean_search_index_db():
     """
@@ -43,11 +38,13 @@ def clean_search_index_db():
             if db.has_collection(name):
                 db.delete_collection(name)
         for name in INDEX_VIEW_NAMES:
-            db.delete_view(name)
+            if name in db.views():
+                db.delete_view(name)
     except CollectionDeleteError as e:
         print("Error deleting collection %s: " % name, e)
     clean_analyzers(db)
     print("search index cleaned.")
+
 
 def clean_all_collections_db():
     """
@@ -59,28 +56,26 @@ def clean_all_collections_db():
         for name in COLLECTION_NAMES:
             current_name = name
             db.delete_collection(name)
-        for name in EDGE_COLLECTION_NAMES:
-            current_name = name
-            db.delete_collection(name)
-        db.delete_graph(GRAPH_COLLECTIONS_CATEGORIES)
     except CollectionDeleteError as e:
         print("Error deleting collection %s: " % current_name, e)
     except GraphDeleteError as e:
         print("couldn't remove graph. It probably doesn't exist.", e)
 
     print("all collections cleaned.")
-    
 
-def clean_totals_collection_db():
+
+def clean_global_stats_db():
     """
-    Clear the categories_parallel_count collection
+    Clear global stats data
     """
     db = get_database()
-    db.delete_collection(COLLECTION_CATEGORIES_PARALLEL_COUNT)
-    db.create_collection(COLLECTION_CATEGORIES_PARALLEL_COUNT)
-    print("totals collection cleaned.")
+    db.delete_collection(GLOBAL_STATS_CATEGORIES)
+    db.create_collection(GLOBAL_STATS_CATEGORIES)
+    db.delete_collection(GLOBAL_STATS_FILES)
+    db.create_collection(GLOBAL_STATS_FILES)
+    print("global stats data cleaned.")
 
-    
+
 def empty_collection(collection_name: str, db: StandardDatabase, edge: bool = False):
     try:
         db.delete_collection(collection_name)
@@ -92,78 +87,74 @@ def empty_collection(collection_name: str, db: StandardDatabase, edge: bool = Fa
     except CollectionCreateError:
         print(f"couldn't create collection: {collection_name}")
 
-        
+
 def clean_segment_collections_db():
     """
     Clear the segment database collections completely.
     """
     db = get_database()
-
-    try:
-        db.delete_graph(GRAPH_FILES_SEGMENTS)
-        db.delete_graph(GRAPH_FILES_PARALLELS)
-        for name in (
-            COLLECTION_SEGMENTS,
-            COLLECTION_PARALLELS,
-            COLLECTION_FILES,
-            COLLECTION_FILES_PARALLEL_COUNT,
-        ):
-            empty_collection(name, db)
-    except (GraphDeleteError, CollectionDeleteError):
-        print(
-            f"couldn't remove graph: {GRAPH_FILES_SEGMENTS}. It probably doesn't exist."
-        )
+    for name in (
+        COLLECTION_SEGMENTS,
+        COLLECTION_PARALLELS,
+        COLLECTION_FILES,
+        COLLECTION_FILES_PARALLEL_COUNT,
+    ):
+        empty_collection(name, db)
     print("segment collections cleaned.")
 
-    
+
 def clean_menu_collections_db():
     """
     Clear the menu database collections completely.
     """
     db = get_database()
-    try:
-        db.delete_graph(GRAPH_COLLECTIONS_CATEGORIES)
-        for name in (
-            COLLECTION_MENU_COLLECTIONS,
-            COLLECTION_MENU_CATEGORIES,
-            COLLECTION_LANGUAGES,
-        ):
-            empty_collection(name, db)
-        for name in (
-            EDGE_COLLECTION_LANGUAGE_HAS_COLLECTIONS,
-            EDGE_COLLECTION_COLLECTION_HAS_CATEGORIES,
-            EDGE_COLLECTION_CATEGORY_HAS_FILES,
-        ):
-            empty_collection(name, db, edge=True)
-    except (GraphDeleteError, CollectionDeleteError):
-        print(
-            f"couldn't remove object {GRAPH_COLLECTIONS_CATEGORIES}. It probably doesn't exist."
-        )
+    db.delete_graph(GRAPH_COLLECTIONS_CATEGORIES)
+    for name in (
+        COLLECTION_MENU_COLLECTIONS,
+        COLLECTION_MENU_CATEGORIES,
+        COLLECTION_LANGUAGES,
+    ):
+        empty_collection(name, db)
     print("menu data collections cleaned.")
 
 
 def clean_all_lang_db(current_lang):
-    print("Cleaning data for language",current_lang)
+    print("Cleaning data for language", current_lang)
     db = get_database()
 
     segments_collection = db.collection(COLLECTION_SEGMENTS)
-    segments_collection.delete_match({"lang":current_lang})
+    segments_collection.delete_match({"lang": current_lang})
 
     parallels_collection = db.collection(COLLECTION_PARALLELS)
-    parallels_collection.delete_match({"src_lang":current_lang})
+    parallels_collection.delete_match({"src_lang": current_lang})
 
     parallels_sorted_collection = db.collection(COLLECTION_PARALLELS_SORTED_BY_FILE)
-    parallels_sorted_collection.delete_match({"lang":current_lang})
+    parallels_sorted_collection.delete_match({"lang": current_lang})
 
     menu_categories_collection = db.collection(COLLECTION_MENU_CATEGORIES)
-    menu_categories_collection.delete_match({"language":current_lang})
+    menu_categories_collection.delete_match({"language": current_lang})
 
     menu_collections_collection = db.collection(COLLECTION_MENU_COLLECTIONS)
-    menu_collections_collection.delete_match({"language":current_lang})
+    menu_collections_collection.delete_match({"language": current_lang})
 
     parallels_count_collection = db.collection(COLLECTION_FILES_PARALLEL_COUNT)
-    parallels_count_collection.delete_match({"language":current_lang})
+    parallels_count_collection.delete_match({"language": current_lang})
 
     files_collection = db.collection(COLLECTION_FILES)
-    files_collection.delete_match({"language":current_lang})
+    files_collection.delete_match({"language": current_lang})
+    print("Cleaning data done.")
+
+def clean_paralels_lang_db(current_lang):
+    print("Cleaning data for language", current_lang)
+    db = get_database()
+
+    parallels_collection = db.collection(COLLECTION_PARALLELS)
+    parallels_collection.delete_match({"src_lang": current_lang})
+
+    parallels_sorted_collection = db.collection(COLLECTION_PARALLELS_SORTED_BY_FILE)
+    parallels_sorted_collection.delete_match({"lang": current_lang})
+
+    parallels_count_collection = db.collection(COLLECTION_FILES_PARALLEL_COUNT)
+    parallels_count_collection.delete_match({"language": current_lang})
+
     print("Cleaning data done.")
